+++ src/app/chat/page.jsx
@@ -4,6 +4,7 @@ import ResponsiveAppBar from '../../components/appbar/ResponsiveAppBar'
 import Chatbox from '../../components/chatbox/Chatbox';
 import React, { useState } from 'react';
 import './chat.scss';
+import { useRouter } from 'next/navigation'
 import SearchIcon from '@mui/icons-material/Search';
 import Searchbar from '../../components/searchbar/Searchbar'
 import ChatMessage from '../../components/chatmessage/ChatMessage'
@@ -14,14 +15,7 @@ import SimpleBottomNavigation from '../../components/navbar/Navbar'
  
 export default function Chat() {
 
-  const [showFirstPage, setShowFirstPage] = useState(true);
-  const [showSecondPage, setShowSecondPage] = useState(false);
-
-  // Funzione per cambiare la visibilità delle classi
-  const handleClick = () => {
-    setShowFirstPage(!showFirstPage);
-    setShowSecondPage(!showSecondPage);
-  };
+  const router = useRouter()
 
   return (
 
@@ -32,13 +26,15 @@ export default function Chat() {
       </ResponsiveAppBar>
 
       <div id="content-panel" className='panel'>
-        {/* Mostra 'firstPage' se showFirstPage è true */}
-        {showFirstPage && (
+
+      <div class="center-box">
+  <h3 class="selectconvo">Seleziona una chat per aprire la conversazione</h3>
+</div>
           <div className='firstPage' id='firstPage'>
             
-            
+         
             {/* Add the onClick handler to make the component clickable */}
-            <div onClick={handleClick} style={{ cursor: 'pointer' }}>
+            <div onClick={() => router.push('/messages')} style={{ cursor: 'pointer' }}> 
             <Chatbox userName="Marco Rossi" messagePreview="Congratulazioni per tutto!" time="16:02" />
             <Chatbox userName="Luigi Verdi" messagePreview="Va bene." time="15:34" />
             <Chatbox userName="Anna Esposito" messagePreview="Ci vediamo dopo." time="15:09" />
@@ -57,40 +53,9 @@ export default function Chat() {
             </div>
 
           </div>
-        )}
 
-        {/* Mostra 'secondPage' se showSecondPage è true */}
-        {showSecondPage && (
-          <div className='secondPage' id='secondPage'>
-          
-          <div className="rightContainer">
-            <ChatMessage className='rightelement' contenuto="Ciao Ugo, devo darti una notizia sul gruppo ambientale che seguiamo..." data="12:02" />
-          </div>
-          <div className="rightContainer">
-            <ChatMessage className='rightelement' contenuto="Non sapevo che avessero creato un evento!" data="12:05" />
-          </div>
 
-          <div className="leftContainer">
-            <ChatRecMessage className='leftelement' contenuto="Me ne sono accorto solo adesso! Anche se non so se riuscirò a partecipare insieme a te. C’è un progetto a cui sto lavorando che mi sta prendendo parecchio tempo..." data="12:12" />
-          </div>
-          
-          <div className="rightContainer">
-            <ChatMessage className='rightelement' contenuto="Ah, capisco... Secondo me dovresti prenderti del tempo libero per rilassarti." data="12:25" />
-          </div>
-
-          <div className="rightContainer">
-            <ChatMessage className='rightelement' contenuto="Sai, ci sono tanti eventi a cui vorrei partecipare insieme a te. Comprendo che tu debba dedicare tempo ed energie all’università, però se riesci a ritagliarti del tempo per andare insieme a questo evento che ti ho menzionato, mi farebbe davvero piacere!" data="12:35" />
-          </div>
-
-          <div className="leftContainer">
-            <ChatRecMessage className='leftelement' contenuto="Va bene, Ugo. Sono contento della tua passione sull’ argomento e di quanto tu ci tenga alla mia compagnia, per cui mi impegnerò per riuscire a partecipare a questo evento!" data="12:42" />
-          </div>
-          
-        </div>
-        
-        
-        )}
-
+      
         </div>
